+++ local_planner/src/nodes/planner_functions.cpp
@@ -125,16 +125,16 @@ void filterPointCloud(
 }
 
 // Calculate FOV. Azimuth angle is wrapped, elevation is not!
-void calculateFOV(double h_fov, double v_fov, std::vector<int> &z_FOV_idx,
-                  int &e_FOV_min, int &e_FOV_max, double yaw, double pitch) {
-  double z_FOV_max =
-      std::round((-yaw * 180.0 / M_PI + h_fov / 2.0 + 270.0) / ALPHA_RES) - 1;
-  double z_FOV_min =
-      std::round((-yaw * 180.0 / M_PI - h_fov / 2.0 + 270.0) / ALPHA_RES) - 1;
+void calculateFOV(float h_fov, float v_fov, std::vector<int> &z_FOV_idx,
+                  int &e_FOV_min, int &e_FOV_max, float yaw, float pitch) {
+  int z_FOV_max =
+      std::round((-yaw * 180.0f / M_PI + h_fov / 2.0f + 270.0f) / ALPHA_RES) - 1;
+  int z_FOV_min =
+      std::round((-yaw * 180.0f / M_PI - h_fov / 2.0f + 270.0f) / ALPHA_RES) - 1;
   e_FOV_max =
-      std::round((-pitch * 180.0 / M_PI + v_fov / 2.0 + 90.0) / ALPHA_RES) - 1;
+      std::round((-pitch * 180.0f / M_PI + v_fov / 2.0f + 90.0f) / ALPHA_RES) - 1;
   e_FOV_min =
-      std::round((-pitch * 180.0 / M_PI - v_fov / 2.0 + 90.0) / ALPHA_RES) - 1;
+      std::round((-pitch * 180.0f / M_PI - v_fov / 2.0f + 90.0f) / ALPHA_RES) - 1;
 
   if (z_FOV_max >= GRID_LENGTH_Z && z_FOV_min >= GRID_LENGTH_Z) {
     z_FOV_max -= GRID_LENGTH_Z;
